Add new method to get folder for Sheets/Drive
By default do not fetch the folders for the Sheets/Drive actions
Provide a way to only fetch on explicit ask, while providing a way
to enter a url of a folder and extract the folder id from it instead.
Solves several issues with large Drives not being able to select all
folders.