Adds checks to ensure error message gets correctly extracted. (#673)
* Adds checks to ensure error message gets correctly extracted.

There have been customer issues where they are receiving an error 'error
on child: {}'. This usually indicates that the error has non-enumerable
properties. This change adds a check to see if the error is of type
Error. It also adds increased logging to tell us if the error cannot be
Stringified.

* Adds checks to ensure error message gets correctly extracted.

There have been customer issues where they are receiving an error 'error
on child: {}'. This usually indicates that the error has non-enumerable
properties. This change adds a check to see if the error is of type
Error. It also adds increased logging to tell us if the error cannot be
Stringified.

* Adds checks to ensure error message gets correctly extracted.

There have been customer issues where they are receiving an error 'error
on child: {}'. This usually indicates that the error has non-enumerable
properties. This change adds a check to see if the error is of type
Error. It also adds increased logging to tell us if the error cannot be
Stringified.

* Adds checks to ensure error message gets correctly extracted.

There have been customer issues where they are receiving an error 'error
on child: {}'. This usually indicates that the error has non-enumerable
properties. This change adds a check to see if the error is of type
Error. It also adds increased logging to tell us if the error cannot be
Stringified.